BJ-969: Fix tests
