+++ src/test/java/com/rbkmoney/wallets_hooker/handler/AdditionalHeadersGeneratorTest.java
@@ -1,7 +1,6 @@
-package com.rbkmoney.wallets_hooker.handler.poller.impl;
+package com.rbkmoney.wallets_hooker.handler;
 
 import com.rbkmoney.wallets_hooker.domain.WebHookModel;
-import com.rbkmoney.wallets_hooker.handler.AdditionalHeadersGenerator;
 import com.rbkmoney.wallets_hooker.service.crypt.AsymSigner;
 import com.rbkmoney.wallets_hooker.service.crypt.KeyPair;
 import org.junit.Test;
