+++ src/test/java/com/rbkmoney/wallets_hooker/handler/WalletEventSinkEventHandlerTest.java
@@ -1,14 +1,9 @@
-package com.rbkmoney.wallets_hooker.handler.poller;
+package com.rbkmoney.wallets_hooker.handler;
 
-import com.rbkmoney.fistful.destination.Authorized;
-import com.rbkmoney.fistful.destination.Status;
-import com.rbkmoney.fistful.destination.StatusChange;
-import com.rbkmoney.fistful.destination.Unauthorized;
-import com.rbkmoney.kafka.common.serialization.ThriftSerializer;
-import com.rbkmoney.machinegun.eventsink.MachineEvent;
-import com.rbkmoney.machinegun.msgpack.Value;
 import com.rbkmoney.wallets_hooker.HookerApplication;
+import com.rbkmoney.wallets_hooker.constant.EventTopic;
 import com.rbkmoney.wallets_hooker.dao.AbstractPostgresIntegrationTest;
+import com.rbkmoney.wallets_hooker.dao.EventLogDao;
 import com.rbkmoney.wallets_hooker.dao.webhook.WebHookDao;
 import com.rbkmoney.wallets_hooker.domain.WebHookModel;
 import com.rbkmoney.wallets_hooker.service.WebHookMessageSenderService;
@@ -26,8 +21,8 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
 
+import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -37,7 +32,7 @@ import static org.mockito.Mockito.verify;
 @SpringBootTest(classes = HookerApplication.class)
 @TestPropertySource(properties = "fistful.pollingEnabled=false")
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class WaitingDestinationAndWalletHandlerTest extends AbstractPostgresIntegrationTest {
+public class WalletEventSinkEventHandlerTest extends AbstractPostgresIntegrationTest {
 
     @Autowired
     private WalletEventService walletEventService;
@@ -51,42 +46,37 @@ public class WaitingDestinationAndWalletHandlerTest extends AbstractPostgresInte
     @Autowired
     private WebHookDao webHookDao;
 
+    @Autowired
+    private EventLogDao eventLogDao;
+
     @MockBean
     private WebHookMessageSenderService webHookMessageSenderService;
 
     @Test
-    public void handleWaitingDestinationAndWallet() throws InterruptedException {
+    public void handle() {
         WebHookModel webhook = TestBeanFactory.createWebhookModel();
 
         webHookDao.create(webhook);
+
         destinationEventService.handleEvents(List.of(TestBeanFactory.createDestination()));
-
         destinationEventService.handleEvents(List.of(TestBeanFactory.createDestinationAccount()));
-
-        MachineEvent destination = TestBeanFactory.createDestination();
-        com.rbkmoney.fistful.destination.Change change = new com.rbkmoney.fistful.destination.Change();
-        change.setStatus(StatusChange.changed(Status.authorized(new Authorized())));
-        destination.setData(Value.bin(new ThriftSerializer<>().serialize("", change)));
-        destinationEventService.handleEvents(List.of(destination));
-
-        change.setStatus(StatusChange.changed(Status.unauthorized(new Unauthorized())));
-        destination.setData(Value.bin(new ThriftSerializer<>().serialize("", change)));
-        destinationEventService.handleEvents(List.of(destination));
-
         walletEventService.handleEvents(List.of(TestBeanFactory.createWalletEvent()));
-
-        CountDownLatch latch = new CountDownLatch(1);
-        new Thread(() -> {
-            MachineEvent event = TestBeanFactory.createWithdrawalSucceeded();
-            withdrawalEventService.handleEvents(List.of(event));
-            latch.countDown();
-        }).start();
-
         withdrawalEventService.handleEvents(List.of(TestBeanFactory.createWithdrawalEvent()));
+
         verify(webHookMessageSenderService, times(1))
                 .send(any());
 
-        latch.await();
-    }
+        withdrawalEventService.handleEvents(List.of(TestBeanFactory.createWithdrawalSucceeded()));
+        verify(webHookMessageSenderService, times(2))
+                .send(any());
 
+        Long lastEventId = eventLogDao.getLastEventId(EventTopic.DESTINATION, 0L);
+        assertEquals(2L, lastEventId.longValue());
+
+        lastEventId = eventLogDao.getLastEventId(EventTopic.WALLET, 0L);
+        assertEquals(TestBeanFactory.WALLET_ID, lastEventId.longValue());
+
+        lastEventId = eventLogDao.getLastEventId(EventTopic.WITHDRAWAL, 0L);
+        assertEquals(67L, lastEventId.longValue());
+    }
 }
